add support for Darwin (macOS)
All the phoenix-2.0 tests now build and run on Darwin targets (macOS), tested on
a Mac Mini with an Apple M1 chip.

Unfortunately, macOS does not support setting CPU affinity for threads. Contrary
to the Linux version, the calls to functions (un)binding threads to cores do
nothing, leaving the kernel scheduler to do this work.

The string_match-pthread test does not build because of a missing header on
macOS.